[refactor] pagination text getter to utils
