+++ Frontend/src/app/components/admin/user-manager/user-manager.component.ts
@@ -1,7 +1,10 @@
 import { Component, effect, inject, signal } from '@angular/core';
 import { AdminService } from '../../../services/admin.service';
 import { User, UserWithRoles } from '../../../models/user';
-import { getPaginatedResult } from '../../../utils/pagination.utils';
+import {
+  getPaginatedResult,
+  getPaginationText,
+} from '../../../utils/pagination.utils';
 import { PaginatedResult } from '../../../models/pagination';
 import { PaginationControlsComponent } from '../../pagination-controls/pagination-controls.component';
 import { SimpleModalComponent } from '../../modal/modal.component';
@@ -45,14 +48,7 @@ export class UserManagerComponent {
   });
 
   get paginationText(): string {
-    const pagination = this.paginatedResult?.pagination;
-    if (!pagination) return '';
-
-    const { itemsPerPage, currentPage, totalItems } = pagination;
-    const start = itemsPerPage * (currentPage - 1) + 1;
-    const end = Math.min(itemsPerPage * currentPage, totalItems);
-
-    return `${start}-${end} of ${pagination.totalItems}`;
+    return getPaginationText(this.paginatedResult);
   }
 
   showModal(user: UserWithRoles) {
